adds field::name to class::PublisherSpec and property::publisher to class::Article; removes file::articles.py and include it to file::source.py
